requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-telegram-bot==20.7
requests==2.32.3
beautifulsoup4==4.12.3
spacy==3.7.6
transformers==4.44.2
sentence-transformers==3.0.1
networkx==3.3
torch==2.4.1
torch-geometric==2.5.3
tsfresh==0.20.3
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2